My-app/observa-store/backend/requirements.txt
# backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-psycopg2
opentelemetry-exporter-otlp-proto-grpc
